refactor into a workspace
