+++ crates/mcdmrs-error/Cargo.toml
@@ -1,0 +1,5 @@
+[package]
+name = "mcdmrs-error"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
