+++ crates/mcdmrs-error/src/lib.rs
@@ -1,0 +1,1 @@
+pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
