+++ crates/mcdmrs-prom/src/math.rs
