+++ crates/mcdmrs-prom/src/types.rs
@@ -4,6 +4,7 @@ pub type Fl = f32;
 pub type Arr = Vec<Fl>;
 pub type Mat = Vec<Arr>;
 pub type FPref = fn(&Fl, &Fl, &Fl) -> Fl;
+
 pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
 pub trait FromVec2 {
